Updated project backend to handle repos without write access.
The project API was not properly handling repos to which the authenticated user
or GH app did not have write access. project.update now does not push what it has
checked out with git itself, but instead, if a reset is being done, calls
pushProjectToGithub, which has been updated to handle resets by doing the
clone/force-push that was previously done in project.update.

pushProjectToGithub now returns if it detects that the repo it's assigned to is
not pushable because of lack of proper credentials.

project.find now checks if projects that the user has permission for are pushable by
the GH app before determining which are tagged as `hadWriteAccess`.