Fixed bug in resetting project to main branch.
Reset was using git.branchLocal, which just lists local branches, and then checking out branchName.
If the remote repo exists, this was checking out the n-commits-behind deployment branch, not the
main branch. Changed this so that on reset, it does git.checkoutLocalBranch, and only on non-resets
does it do git.checkout