Fixed issues with key stuck on instance load and chat getting movement keys.
Having 'autofocus' on the instance chat was causing problems. Since it's not display:none
when closed, and it doesn't load until the instance has been connected to, if it was rendered when
the user was moving, it was now getting focus. Their movement key presses were getting entered into
the instance chat, and movement was never seeing the keyup that would tell it to stop. Removed
autofocus from instance chat.