Added error handling for instance provisioning returning a shutting-down instance
There are corner cases where an instance-provision occurs right before an instanceserver
shuts down, or has its shutdown status recorded in the database. The provision returns an
IP address that is immediately invalid. The SocketWebRTCClientNetwork instantiation was not
handling this case, and throwing errors.

Now, whether there's an error with creating the websocket connection, or it fails to connect
within 3 seconds, it disconnects and then attempts another instance-provision.