Entity Tree Improvements (#6993)
* various scene fixes

* fix entity tree bugs

* test fixes

* Fixed issues with key stuck on instance load and chat getting movement keys. (#6994)

Having 'autofocus' on the instance chat was causing problems. Since it's not display:none
when closed, and it doesn't load until the instance has been connected to, if it was rendered when
the user was moving, it was now getting focus. Their movement key presses were getting entered into
the instance chat, and movement was never seeing the keyup that would tell it to stop. Removed
autofocus from instance chat.

* Lightmap and texture editor fixes (#6995)

* add scale and offset fields in texture editor. improved thumbnail handling

* create static renderer & scene for createReadableTexture instead of initializing and disposing of them every function call

* Fixed issue with redis connection string password. (#6997)

feathers-sync v3 uses redis v4, which requires that the password be passed at the start of the
connection string, as opposed to as a query parameter. Made this change.

* fix editor tests

* fix asset component

Co-authored-by: Gheric Speiginer <gheric.speiginer@gmail.com>
Co-authored-by: Kyle Baran <kbaran@bitscoop.com>
Co-authored-by: dinomut1 <94419856+dinomut1@users.noreply.github.com>
Co-authored-by: dinomut1 <heximhotep@gmail.com>