Added support for project versioning, better project updating. (#7046)
Projects' package.json's should now include a section `etherealEngine`, which
contains the `version` of the engine that it is known to be compatible with.
The name field is now the definitive name of the project - the name of the
repository that the project is in is not used for any naming purposes.
The version in here is also displayed, and tags are used for retrieving
specific commits.

Replaced Update and Reset functionality with a single drawer. This drawer lets
user select a destination repository, as well as a source repository/branch/tagged
commit (the latest commit is also selectable). The API checks that the source and
destination are accessible by the user, and that the user has push permissions for
the latter. It also checks that the destination is either empty, or contains a
project whose name matches the source project's name.

A warning is displayed if the source project's engine version does not match
the deployment's project version, but it is always accepted. Update now always
sends `reset: true` to project.update, so whatever is in the
<deployment>-deployment branch in the destination is overwritten.

Updated builder update to a) allow for choosing which commit in the builder's repo
to use when doing the update and b) select projects to update in the new builder
update drawer.