Fixed bug in pruning old S3 client files.
Vite sometimes generates the same hash for a built file as the prior run. These files were getting deleted
at the end of the build process because they were being recorded before the client build as files that
existed and should be deleted, leading to missing files at the end of the process.

Made get-deletable-client-files.ts store the list of files to be pruned stored in S3. push-client-to-s3.ts
now reads in the list from S3 and removes files whose names are the same as the just-built client files,
then pushes that back to S3. delete-old-s3-files.ts gets the list from S3.