Fixed several bugs in project versioning.
project.find no longer errors out if any individual project fetch from Github
fails. The returned error is caught and placed into an array of errors.
This array is returned alongside the valid data, to be displayed to the requesting
user.

project.find's call to getRepo was not removing the `.git` from the repository name,
which was resulting in invalid calls. It now is.

Failing to get a project's package.json is now caught and ignored. The information
that is expected to be retrieved is just not returned.

Added GitHub scopes to authentication seed. These are necessary for full access to
repos for projects.

Removed a redundant fetch of projects on admin projects page.

Added a button in ProjectFields.tsx to copy the destinationURL to the sourceURL.

Reject project sources that lack an xrengine.config,ts