Fixed several avatar-related bugs
Spawning of avatar now uses avatar object on user object. Client had previously
been using fetchAvatarList and pulling the matching avatar off that, which was
prone to failure if it was not on the first page. Instanceserver now gets user
via feathers instead of sequelize model so that avatar is attached to user, and
does not have to do a separate fetch of user's avatar.

Made project.update not ru onInstall hook if the project already exists in the database.
This was causing avatars to be duplicated.

Updated fetchAvatarList to paginate. Updated avatar selection menus to paginate themselves
by their page length and to make calls to fetchAvatarList when they need to get more from
the API.