Updated instance-provisioning to support private instances. (#7159)
* Updated instance-provisioning to support private instances.

Changed input query param 'createNewRoom' to 'createPrivateRoom'. If this is passed, then it mostly
works as before, forcing a new instance to be provisioned, but it will also make the requesting user
be added as an instance_authorized_user. This will prevent ordinary instance provision requests from
putting users on this instance.

Changed params in most instance-provision.class.ts functions to be objects, as a long list of input variables
when many of them are optional becomes unwieldy. If a user provisions with a room code and that instance exists,
an instance_authorized_user will be made for them if it does not exist.

Changed default domain on minikube build from theoverlay.io to etherealengine.com

* fixup