Added configuration of site manifest and automatic reupload of manifest.
With the move to hosting client files from the storage provider, the site manifest file was
providing improper paths. The start_path was being read by the client as being under the
storage provider's domain, rather than the client domain, and the icon paths were not correct,
since the convention is to have client files under /client in the storage provider.

Most aspects of the site manifest are now configurable from clientSettings. 'name' uses
clientSettings.title; 'short_name' comes from a new field, clientSettings.shortTitle;
'start_url' explicitly adds a new field clientSettings.startPath onto config.client.url;
and the two icons are combined from clientSettings.icon192px and clientSettings.icon512px
with the cacheDomain.

client-setting.patch now pulls down the manifest from the storage provider, updates it with the
patched data, and pushes it back to the storage provider. The same basic thing is done during
run-builder.sh in a new script update-site-manifest.ts, but instead of pulling/pushing to the
storage provider, just updates the copy of the file in packages/client/public, as all of that
will be built and pushed to the storage provider in the client build process.