Fixed some issues with mobile styling, particularly homepage.
Homepage was stretching beyond edge of screen on mobile devices. Fixed this so that main contents
scroll, while header and social footer are fixed.

Editor Projects Page now scrolls on mobile as well.

Made Other Audio Settings selector be the row the button is in rather than just the arrow.