Switched from socket.io to Primus
Since primus uses a hard-coded path for its client library, the old way
of passing the address and port of an instanceserver, via the path,
did not work. They are now passed as query params and read as arg_address
and arg_port in the nginx ingress.

The major change from socket.io is that socket.io could listen for specific
events, while primus has a single 'data' event. This means setting up listeners
for 'data' and, since all events are coming across the same socket, only acting
on specific messages.

promisedRequest required some workarounds. It now adds an ID to each request
so that it can identify the specific response it's looking for. It makes a new
'data' listener with each request, and changes the name to the listener to
include the id of the request. When it receives the response to the request, it
destroys that listener. Without making unique IDs for each listener function,
primus' removeListener function can't tell them apart and destroys all of them.

There's a new address format for instanceserver connections. Before, we were adding
the IP and port on the path, e.g.
instanceserver.etherealengine.com/socket.io/255.255.255.255/7999?<querystring>
Primus' client has a hard-coded path that it tries to reach, and will ignore anything
in the path given to it. The IP and port are now added to the querystring, and the
nginx ingress is changed to read those querystring arguments and replace the address
with them, e.g.
instanceserver.etherealengine.com/primus?address=255.255.255.255&port=7999&<etc>