Fixed multiple issues with projects.
Install project button on editor project page wasn't working because
page assumed only an active project would make the drawer be open.
Now ProjectDrawer renders all the time.

When fetching project commits, commits that didn't have a package.json
were throwing errors, which would cause the whole request to error.
This was not useful behavior, so now commits without a package.json
are discarded.

Errors on commit fetching after selecting a branch were setting the wrong
error, which was causing those errors to not be displayed.

Made selecting a commit from the menu blur the input, so that you don't have
to click again to blur it and trigger the commit check.