Added storage of build status and build error logs (#7354)
A new table, build_status, stores the status of every run of the builder.
On builder errors, the logs from the build process that failed are stored
in this table as well.

Added a drawer to /admin/projects to retrieve and view build statuses in
reverse chronological order, and a drawer that opens with failure logs.

Updated run_builder.sh to work with this. Near the start of the build
process, it creates a build record, and sets all builds that are still
in the 'pending' state to 'ended' - these are builds that restarted
without an obvious error, like a new build being started a couple
minutes later, or random restarts that occasionally occur.
Most commands that are run have their stdout and stderr logs output
to files, and if there's an error detected, a script for recording
errors is called. Docker build error detection is done different
from other since BuildKit writes all logs to stderr, meaning we
have to check for the presence of 'ERROR:' to see if there was
a failure. If deployment is successful, then a third script is called
to mark that build as successful.

Made download-projects skip all files in the top-level /assets folder
of a project. This folder should be for static assets that do not
need to be built, and should only be served from the storage provider.
This can dramatically cut down on the final Docker image size.