Updated installing avatars from projects.
avatar and static-resource tables now have a column to link them to a project.
When a installAvatarsFromProject is called, it makes new avatars and their
associated static-resources linked to that project.

installAvatarsFromProject also now checks if each avatar with that name, and associated
with that project, exists before inserting it. This allows avatars to be added or changed
without creating duplicates.

When a project is deleted, it now deletes all avatars and static-resources associated
with that project.

Made install-projects call onUpdate on default-project