Implemented auto-updating of projects from source repository.
When a project is added or updated, the repo&branch&commit that are selected are saved in
the project table. Also saved, optionally, is whether to auto-update that project when the
source branch is updated. If enabled, auto-update can be in 'prod' mode, where it only updates
to the most recent tagged commit, or 'dev' mode, where it will update to any new commits.
The timing of how often to check for updates is also configurable.

This is implemented as creating a Kubernetes CronJob with the selected interval. This CronJob
will run a script that checks if there's an eligible new commit and, if so, runs project.update
with the new comit.

The builder runs an additional script after deploying that updates all projects' CronJobs with
the new API image tag, as that is the image that the CronJob uses.

/admin/server now also shows pods created by project CronJobs.