Refactored how assets/static resources work.
Most types of assets (videos, volumetric, models, etc.) now have their own table.
This table links to one or more static resources, possibly through other asset tables.
e.g. volumetric links directly to a static resource for the .drcs/.uvol file, and also
links to a video for the .mp4 texture, and a data for the .manifest file.

Static resources now have 18 sets of columns for Levels of Detail and associated metadata.

When an asset is created, either through the studio or through directly uploading via
/admin/resources, a top-level asset is created, along with the associated static-resource.
The file(s) is downloaded from the source URL and uploaded to the storage provider; the URL
in the storage provider is what's saved on the static-resource. When a scene is saved, it
will now save the entirety of the asset and its linked assets and resources. All links
should be to copies stored in the deplyoment's storage provider, not external sources.

A new script will convert static_resource.url to static_resource.LOD0_url, so that
existing resources will work. Front-end code does have fallbacks for old resources.

When adding a volumetric video through the studio, one must now point to either the
.drcs file or the .manifest, but not the .mp4. This is due to current lack of information
on the MediaNodeEditor about what types of assets it's handling; it assumes mp4s are
videos, and will auto-upload accordingly.