Fixed bugs in project auto-update.
Combined creating and patching project cronjob into one function. It tries
to patch the named cronjob, and if it doesn't exist, tries to create that
cronjob instead.

Made update-cronjob-image script try/catch its patching of existing cronjobs
so that failures don't throw an error and restart the builder process.

Added copying of /patches to local Dockerfile