Added checks to deploy.sh to prevent errant helm upgrades. (#7450)
Sometimes, several failed builds in a row are triggering the helm upgrade in deploy.sh.
Added checks to try to prevent this from happening. deploy now calls `docker manifest inspect`
on each image that should have been built, pipes the output to a file, and only does the helm
upgrade if all expected files are present and do not contain 'no such manifest', which is what
is returned if the image is missing.