Added support for storing Github repo access in the database.
Loading projects with associated push access was slow because the backend process
was making many requests to the GitHub API each time. Now, which repos a user
has access to is stored in a new database table. When a user logs in with GitHub,
their list of repo access is updated. Both /admin/projects and /studio have a button
that users logged in with GitHub can click to trigger the same behavior. Also added
a route for GH webhooks to trigger this re-fetch when a user is added/modified/removed
as a collaborator on a repo or an organization. Support for changes to teams is not
included because getting the list of users in a team is tricky.

Fixed an issue with users not being able to update a project from GitHub if they were
not part of the organization that owned the repo. This was leftover thinking from GitHub
Apps, which required Org access, and instead the backend check now gets the users allowed
repos instead.

Replaced some useState's with useHookstate's.

Updated the studio projects page to determine display of most projects' actions via the
same method as the admin projects page.