Fixed bugs with GH repo access
Project update destination check was not comparing the correct URL string.

Added 'hasWriteAccess' to github-repo-access model. This is needed to
distinguish whether a user can push to that repo. github-repo-access-refresh now
records this based on whether a user has admin, maintain, or write access to the
repo. github-repo-access entries are now patched if they exist, to account for
changes in write permission.

Updated allowed project find to better express push access.