Fixed some missing app.setup()'s in scripts.
The install-projects script was not calling app.setup() after creating the feathers
app. With the new creation of assets when default-project is uploaded, this was
causing the resulting find's with associations to fail because the associations
were not set up.

Saw a couple other scripts that had this problem as well, added there too.

Added some legacy fallback for avatar urls.