Pwabuilder domain setting (#7767)
* Added client setting for pwabuilder-sw.js.

This was needed to make it configurable, so that if client files are
served from the storage provider, the service worker is fetched from there
and not the app subdomain, as it might go through a 304 on index.html and
not be seen as the proper file type.

Converted a lot of deprecated use of state. Removed access<Service>State
and use<Service>State usage for all admin services. All files that were
using these deprecated state usage now access is via useHookstate.
In all of these files, also convered all usage of React's useState to
useHookstate.

Most admin pages have been converted, though some that were not using
any service's state, just useState, were omitted. Non-admin pages
and components that were using admin service in the deprecated fashion
have been converted, though there are a number that have not.

Fixed some bugs with OAuth provider icon sizes.

* Fix component toJSON

* Fix type errors

---------

Co-authored-by: Gheric Speiginer <gheric.speiginer@gmail.com>