Fixed some bugs with ProfileMenu and InputText cursor. (#7874)
ProfileMenu had reverted to not showing single OAuth identity provider. Fixed this.

Proper setting of cursor in InputText had been broken by recent changes. Fixed this
by making a default inputRef and using this to set the selectionRange.

Made some slices of commitSHA's optional in case they somehow don't exist, lest
the client page crash as a result (I'd seen this on a deployment, and it prevents
updating the engine version).