Fixes a client-side bug with users' avatars missing resources.
Sometimes, a user-patched message is getting sent to the client
where the nested avatar object does not have modelResource or
thumbnailResource. This was causing errors due to assuming
modelResource would always exist.

Workaround is to not assume that, and to attempt to fetch the
avatar manually if it is missing modelResource. It will then
spawn the avatar from this, and only if this failed to get the
necessary information is an error thrown about the avatar
missing its model.