Fixed and improved Docker build caching. (#7988)
Builder build process caching was not working because it was a multi-stage build.
The default inline cache only caches the final stage of a multi-stage build, which
almost always results in cache misses on the earlier stages. All builds in the builder
were not being cached at all.

Converted builds to explicitly use buildx. buildx build has support for automatically
pushing tags, so separate steps for tagging and pushing images have been removed.

In order to cache multi-stage builds, buildx has an option to cache everything. This
has to be done with a cache target of something other than inline cache (which includes
cache in the image). There is an option to use a Docker repository, but currently this
is not supported by ECR. For now, using S3 as the cache target (another built-in option)
for service images, and GH Actions cache for builder image.
An issue has been opened to switch to the repo for the service images once it's supported.

Commented out running npm install when each project is installed. This appeared to be causing
issues where some packages, particularly semver, were not fully installed properly by the
end, leading to errors when those packages were referenced. There is now a single npm
install after installing all projects.