Fixed some bugs in cache flow
Running npm install from packages/projects was not correct. To just install
projects' dependencies, we're now using a second root package.json that for
workspaces only has the packages that the builder comes preinstalled with
and packages/projects/projects/*. This ensures that all projects'
dependencies are installed, all necessary packages for running the builder
stay installed, and packages that are not necessary for root builder tasks
are not installed unnecessarily.

Dockerfile-root runs npm install; this had been commented out for testing
purposes and mistakenly not uncommented.