Allows for serving client from API pods. (#7351)
Added configurable ability to build and serve client files from
API pods instead of standalone client pods. Added a new API
Dockerfile that builds the API and client portions in separate
stages, then copies just the built client files into the final
image.

build_and_publish_package.sh now takes an additional parameter
for the Dockerfile name, so that there can be multiple Dockerfiles
for the same service. In this case, the api server can be built
plainly with Dockerfile-api, or with the client included with
Dockerfile-api-client.