Fixed a bug with failed connection logic. (#8033)
When a connection fails, it will sometimes provision a new server.
This was only checking if the current connection's 'connected' status
was true, but not whether the 'connecting' status was true. On some
long-loading scenes, the initial connection was failing right as a
second provision was starting, and was leading to two connections
being made concurrently, which would cause problems - the second
connection would not make any dataProducers/dataConsumers, leaving
the user with no networking despite being connected.

Fixed a few bugs that were causing type errors.