Added logic to prevent reconnection warning when switching media servers
MediaInstanceConnectionService now has a field 'joiningNewMediaChannel'.
This is set to true when actively switching from one media server to
another, and to false once connected to that server.

Media disconnection actions will not display the warning modal
if joiningNewMediaChannel is true, as they were disconnected for a
non-error reason.

Added an indicator that media connection is still in progress.