Added logic to avoid Mediasoup using ports in use (#8108)
Mediasoup will throw an error if it tries to run
createWebRtcServer on a port that is in use. This is most
often encountered in local development when things like
IDEs use a port in the 40000 range.

Added some logic that checks if a port is in use before
attempting to run on it. If it is in use, that port is skipped.