Updated client to prefer lowest-res video when video is small. (#8225)
Rendering 1080p video for every user with their camera on can be taxing on a user's
device when several people have their cameras on. Additionally, it's very hard to
tell the difference in quality when the video is so small. Made consumers prefer the
lowest spatial layer (resolution) unless PiP is on, at which point it switches to
the highest spatial layer, or if immersive media is enabled, which will use either
the top layer-1 if there are more than 2, or 0 if there are not.

Updated user window order so that the user's own screen/cam come at the top.

Replaced tcp-port-detect with detect-port, as the former was not detecting certain
processes that were actively running.