Added ability to set Helm chart versions. (#8205)
There is a new settings table, helm-setting. It stores the version
of the main and builder Helm charts that should be used when running
deployments. Build scripts have been updated to reflect this.

Updating the Engine version now uses helm upgrade on the builder deployment
instead of using kubectl to patch the builder configmap. It will use the
version of the builder Helm chart configured in helm settings unless
one is not set, in which case it will use the current builder version.
This requires the main deployment to be on version 5.1.3 or higher, as that
added required cluster roles to the API service to run helm upgrade.

The builder deploys the main deplyoment with the Helm version specified in
the helm settings, unless one is not specified, in which case it uses the
current version.

Added hash navigation to settings sub-pages. Update builder modal now shows
which Helm chart versions will be used, and has links to helm settings
subpage for changing them.