Updated builder to be run as a job instead of a deployment.
This allows the builder to stop if it fails too many times in a row.
It also allows the builder to be scaled better.

Modified the builder status so that what's returned is a succeeded or
failed status, not just whether it is running or not. Added a visual
indicator to /admin/projects of the builder's status (green dot for
success, yellow for running, red for failed).