Fixed a bug where in-use service images could be pruned by failing builds (#8250)
If a service's build ran to completion but another one failed, the successful
service would upload and prune images without deploying the new image. If the
builder failed too many times in a row, this could lead to the image that that
service was still using being pruned away.

Created a new prune_ecr_images.ts for build_and_publish_package.sh to use.
This checks which images are in use by existing pods of that service and omits
them from the list of prunable images.

Updated criteria for which images to save. Previously, it was just the most
recent 10 images that weren't the latest. Due to the way ECR handles multi-
architecture builds, this could remove some images that were tied to others,
particularly with in-use images. Now, it finds all images that were uploaded
within 10 seconds of a tagged image - these are going to be the related images,
with the tagged one being the the image index file. Downloading and parsing the
image index was likely going to be too time-intensive for the benefit.

Added TS scripts to 'format' npm command.