Updated serving client from storage provider. (#8060)
Serving the client from the storage provider now serves everything from there. Previously,
the root index.html file was coming from the EKS cluster (api pod or client pod), while every
other client file was built to and served from the Cloudfront domain. This was inefficient
in multiple ways, both in building and storing client files on Docker image and pods just
to serve that one file, and because it requires maintaining multiple domains for client files.

Client build will now use an alternate Dockerfile that doesn't do anything besides push the
files to S3. build_and_publish_package.sh will not push this image to ECR since it will not
be served by Kubernetes.

All static files will come from the client subdomain, which should be attached to the Cloudfront
distribution instead of the `resources.<domain>` subdomain (a certificate for the client subdomain
is needed in us-east-1 in this case). The only exception is if STORAGE_S3_ENDPOINT is defined,
as this serves from the S3 URL directly instead of through Cloudfront.

If both SERVE_CLIENT_FROM_API and SERVE_CLIENT_FROM_STORAGE_PROVIDER are true, it will serve
from the storage provider, and will not build the client files into the API.

Added a bit to check-db-exists to update the redis credentials. This solves a problem where
a cluster's redis deployment has been removed and added, which results in a different address
and password; the builder should now make sure that it keeps redis information up-to-date.
This should be done here instead of prepare-database since prepare-database tries to connect
to redis, and will fail if the credentials not correct.