Fixed a race condition in ProjectDrawer.
The useEffect on `open`, which would set the project destination field to kick off
auto-population, was triggering before the call to initializeProjectFields in
ProjectFields.tsx's initial render useEffect. This was resulting in the destination
field getting reset to an empty string, and no auto-population was occurring.

Now the `open` useEffect also listens to projectUpdateState.triggerSetDestination,
so that it will fire after the initialization occurs, and will only fire if the
current value is an empty string.