Added helm repo update to updateBuilder.
An overlooked issue was that the helm repo status is set when the builder
builds the API pods. If a new version of an EE-builder Helm chart is made after that,
then trying to update the builder with the new version will throw an error since the
API doesn't know about the new version. Running helm repo update right before helm upgrade
fixes this.