Updated instance-authorized-user migration. (#8412)
In the current state where not everything has been converted off of sequelize,
some sequelize BelongsToMany (join table) associations are bugging because they
are being made under new names without all of the required fields before the knex
migration can create it properly.

Specifically, instance_authorized_user was renamed to instance-authorized-user.
The sequelize BelongsToMany relationship between user and instance is making a
new instance-authorized-user table just based on that relationship, which doesn't
know about the 'id' field. Then the knex migration tries to run, and fails because
the instance-authorized-user table already exists. The deployment then has bugs
because the table schema isn't correct.

The solution was to make the migration delete the new table if it exists if it
detects that the old table still exists. This probably needs to be done for other
join table/BelongsToMany conversions.