Fixed paginate: false for routes (#8436)
If the client passes paginate: false on params, it will not make it to
the API server's handler for security reasons. Changed the find call to
accept paginate: false as a query param, and to move it to the top of the
params object if it's false.

Had to make changes to route's typing to support this. Made some changes
to RouterService to handle the fact that the route.find call now returns
an array rather than a paginated response. Cast several calls to route.find
as either paginated or unpaginated based on what they are requesting.