Switched from caching in S3 to caching in registry
BuildKit now supports caching in ECR. build_and_publish_package.sh now
uses BuildKit 12.0 and caches to and from the registry. Cache image is
tagged as latest_<release>_cache. prune_ecr_images.js updated to ignore
this image.

Fixed an issue with record-build-error.ts If there was a miss on the
build cache and another error occurred, the logic was ignoring errors
and returning as if everything had succeeded. Now checking if exit code
1 was part of the output logs to better catch legitimate errors.