Fixed a few issues with admin avatar drawer. (#8470)
Newly-created avatars did not have thumbnails/models show up until a page refresh
or another avatar was created. This was due to the switch to FeathersHooks; the avatar
is created but did not have the model or thumbnail yet, then it uploads the model and
thumbnail, but useFind() was only being triggered by the creation, not the resources
getting added. Added a patched handler on avatar.ts that will notify people with that
avatar and the user who made the request. Passing params to avatar.patch in
uploadAvatarStaticResource to make this happen.

Thumbnail preview was not showing in AvatarDrawer in create mode because it was set
to only show in edit mode. Removed this restriction, as it is helpful to show the
thumbnail during creation too.