Quick Fix for scene switching bug in studio
This is a quick fix how we are mishandling scene entities that exist in multiple scenes (e.g., the same ground plane UUID when switching scenes in the studio).

Our scene loading state and SceneAssetPendingTagComponent are too ambiguous, prone to multiple sources of truth conflicting with one another. We need to refactor this.