Updated uploadToRepo to use git LFS (#8577)
Existing code did not use git LFS, which made it unusable with projects that had
files too big to upload directly to GitHub, as the blob uploads would be rejected.
createBlobForFile now checks the size of the file and, if it's over 30 MB, which is
low enough to not come close to GitHub's stated 50 MB cutoff,
it instead uploads the file to GitHub's LFS and makes the file a pointer to it instead
(which is standard LFS practice). The flow also adds the files to the .gitattributes
file, creating and adding it to the git history if that doesn't already exist.

Made 'archiver', 'project-push-github', and 'project.update' handlers create jobs
for their respective logic if in a deployed environment. This uses a new standardized
function 'createExecutorJob' in project-helper.ts, in case any other functions should
be run as jobs as well.