Fix for project files using S3 URL
Sometimes files added to projects would have an S3 URL. This seemed to be
occurring when adding files from the file browser. The cause was that the image
node was using the URL from the file browser. These URLs were coming from
storageProvider.get, which for S3 returns the S3 URL.