Use multipart upload on large files (#8667)
* Use multipart upload on large files

While S3's documentation says that putObject should work on files
up to 5 GB, in practice problems were occurring with archive files
that were around 2 GB. Updated the S3 storage provider putObject
function to perform a multipart upload on files >= 1 GB.

* Cleaned nested if conditions

* Use read/write streams for large files locally.

* Fix for archiving when multiple projects share start of name

Need to keep the '/' at the end of the project directory to not match e.g.
test-project-2 when archiving test-project, as the former will be returned
from listFolderContent if there's no slash at the end.

---------

Co-authored-by: hanzlamateen <hanzlamateen@live.com>