Fixed some bugs with the file browser and project paths.
React-DND does not populate a type for GLBs. This was causing it to make
them as folders, which was not correct. Could not figure out where the type
was being populated from, so added a hard check on the file name to avoid
this.

Large file check on S3 putObject was not accounting for being handed a path
with no Body. Made that Body.length check optional to avoid errors.
Made that on local putObject as well just in case.

downloadProjects was tying to write files even if their Body had no length,
i.e. was a folder, which would error out.