Fixed bugs in updateBuilder()
Updating builder when Helm version was not set was bugged. It was using the raw call of
BUIDLER_CHART_REGEX.exec(stdout), when it should have been using item 1 from the result,
which is the actual capture group value, i.e. the version itself.

The above was introducing a buggy state when it failed. The builder job had been deleted,
but the upgrade had not succeeded and created a new job. On subsequent calls, since there
was no job, it was assuming that the builder was a deployment, and then failing to delete
and upgrade the nonexistent deployment. To fix, the deletion of the job/deployment is its
own step, and then it performs a common helm upgrade depending on whether or not the version
is set in helmSettings.