Fixed bugs in uploading assets
isAssetFromProject now checks if the originURL matches the url it's checking
to handle raw S3 URLs. getKeyForAsset now replaces origin URLs.

file-browser.patch was always making a new asset with no associated project.
It now checks if the asset exists and new assets are made associated with
the project they're in, if they are coming from a project.

Working around some issues with using KTX2Loader on the backend.