Fixed issues with static resource hash generation and ktx2 size fetching
If fetching a ktx2 file for the purposes of getting stats, only fetching the
first 28 bytes, as the size is on bytes 20-28. Would have fetched just those
bytes, but it's easier to get the preceding bytes as well since if the file
passed to getImageStats is a buffer, it doesn't truncate the preceding bytes.

Resolved discrepancies in how file hashes for static resources were being generated.
startMediaRecording was only using the file's name, createStaticResourceHash was
using the length and name but not mimeType, and getFileMetadata was using all three.
All methods are now using length, name, and mimeType to generate hashes.