Added some checks for resolvers after removes
Avatar removal was returning errors. This was caused by the avatarResolver fetching
model and thumbnail static-resources, but on avatar deletion, these have been removed,
which makes the .get's error. Made this not trigger when context.event is 'removed',
and applied this to other resolvers that fetch static-resources.

Cleaned up some redundant function calls that are assigned to variables before just
returning the variable.